refactor/#198: BubbleConfiguration 파일 분리
